--- src/main/java/com/example/FixLog/controller/MemberController.java
@@ -1,4 +1,0 @@
-package com.example.FixLog.controller;
-
-public class MemberController {
-}
