--- src/main/java/com/example/FixLog/service/MemberService.java
@@ -1,4 +1,0 @@
-package com.example.FixLog.service;
-
-public class MemberService {
-}
