+++ src/main/java/com/example/fixLog/service/MemberService.java
@@ -1,0 +1,4 @@
+package com.example.fixLog.service;
+
+public class MemberService {
+}
