+++ src/main/java/com/example/fixlog/domain/bookmark/Bookmark.java
@@ -1,7 +1,7 @@
-package com.example.FixLog.domain.bookmark;
+package com.example.fixLog.domain.bookmark;
 
-import com.example.FixLog.domain.member.Member;
-import com.example.FixLog.domain.post.Post;
+import com.example.fixLog.domain.member.Member;
+import com.example.fixLog.domain.post.Post;
 
 import jakarta.persistence.*;
 import lombok.AccessLevel;
