+++ src/main/java/com/example/fixlog/domain/follow/Follow.java
@@ -1,6 +1,6 @@
-package com.example.FixLog.domain.follow;
+package com.example.fixLog.domain.follow;
 
-import com.example.FixLog.domain.member.Member;
+import com.example.fixLog.domain.member.Member;
 import jakarta.persistence.*;
 import lombok.AccessLevel;
 import lombok.Getter;
