+++ src/main/java/com/example/fixlog/domain/post/Post.java
@@ -1,6 +1,6 @@
-package com.example.FixLog.domain.post;
+package com.example.fixLog.domain.post;
 
-import com.example.FixLog.domain.member.Member;
+import com.example.fixLog.domain.member.Member;
 import jakarta.persistence.*;
 import lombok.AccessLevel;
 import lombok.Getter;
