+++ src/main/java/com/example/fixlog/domain/tag/TagCategory.java
@@ -1,4 +1,4 @@
-package com.example.FixLog.domain.tag;
+package com.example.fixLog.domain.tag;
 
 public enum TagCategory {
     STYLE,
