+++ src/test/java/com/example/fixLog/FixLogApplicationTests.java
@@ -1,4 +1,4 @@
-package com.example.FixLog;
+package com.example.fixLog;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
